+++ src/lib/nav/syncStatus/SyncStatusProgress.svelte
